refactor(r2c1): extract live DataExchange into Rhino-free Progesi.LiveDataExchange
R2-C.1 (DataExchange Option D). Moves the live Excel/SQLite export+import logic out of the GH ProgesiDataExchangeComponent into a new Rhino-free library Progesi.LiveDataExchange (built on Progesi.GhExcelReadContract + ProgesiCore). The component becomes a thin wrapper: reads the Rhino StringTable into a LiveExchangeSnapshot and delegates, supplying a RhinoGeometryValueCodecAdapter (IGeometryValueCodec) and RhinoLiveExchangeImportSink (ILiveExchangeImportSink) for Rhino-side persistence. EF path unchanged (shells to Progesi.EF.Tool.exe with SQLite fallback). New unit tests: Excel + SQLite round-trip with a fake geometry codec + in-memory sink. No ProgesiCore change; AxisVar frozen/untouched. Co-Authored-By: Claude Opus 4.8 <noreply@anthropic.com>
