+++ Progesi.sln
@@ -49,6 +49,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Progesi.GhExcelReadContract
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Progesi.Stress.Tests", "tests\Progesi.Stress.Tests\Progesi.Stress.Tests.csproj", "{52C0B54B-9C1A-4D4F-B7F3-99CE3960AE95}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Progesi.LiveDataExchange", "src\Progesi.LiveDataExchange\Progesi.LiveDataExchange.csproj", "{27C83E37-2B9D-4B47-9096-A503C0D53F87}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Progesi.LiveDataExchange.Tests", "tests\Progesi.LiveDataExchange.Tests\Progesi.LiveDataExchange.Tests.csproj", "{F3ECBDF1-3AAD-418B-A005-E9435F5041C5}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -299,6 +303,30 @@ Global
 		{52C0B54B-9C1A-4D4F-B7F3-99CE3960AE95}.Release|x64.Build.0 = Release|Any CPU
 		{52C0B54B-9C1A-4D4F-B7F3-99CE3960AE95}.Release|x86.ActiveCfg = Release|Any CPU
 		{52C0B54B-9C1A-4D4F-B7F3-99CE3960AE95}.Release|x86.Build.0 = Release|Any CPU
+		{27C83E37-2B9D-4B47-9096-A503C0D53F87}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{27C83E37-2B9D-4B47-9096-A503C0D53F87}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{27C83E37-2B9D-4B47-9096-A503C0D53F87}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{27C83E37-2B9D-4B47-9096-A503C0D53F87}.Debug|x64.Build.0 = Debug|Any CPU
+		{27C83E37-2B9D-4B47-9096-A503C0D53F87}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{27C83E37-2B9D-4B47-9096-A503C0D53F87}.Debug|x86.Build.0 = Debug|Any CPU
+		{27C83E37-2B9D-4B47-9096-A503C0D53F87}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{27C83E37-2B9D-4B47-9096-A503C0D53F87}.Release|Any CPU.Build.0 = Release|Any CPU
+		{27C83E37-2B9D-4B47-9096-A503C0D53F87}.Release|x64.ActiveCfg = Release|Any CPU
+		{27C83E37-2B9D-4B47-9096-A503C0D53F87}.Release|x64.Build.0 = Release|Any CPU
+		{27C83E37-2B9D-4B47-9096-A503C0D53F87}.Release|x86.ActiveCfg = Release|Any CPU
+		{27C83E37-2B9D-4B47-9096-A503C0D53F87}.Release|x86.Build.0 = Release|Any CPU
+		{F3ECBDF1-3AAD-418B-A005-E9435F5041C5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F3ECBDF1-3AAD-418B-A005-E9435F5041C5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F3ECBDF1-3AAD-418B-A005-E9435F5041C5}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{F3ECBDF1-3AAD-418B-A005-E9435F5041C5}.Debug|x64.Build.0 = Debug|Any CPU
+		{F3ECBDF1-3AAD-418B-A005-E9435F5041C5}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{F3ECBDF1-3AAD-418B-A005-E9435F5041C5}.Debug|x86.Build.0 = Debug|Any CPU
+		{F3ECBDF1-3AAD-418B-A005-E9435F5041C5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F3ECBDF1-3AAD-418B-A005-E9435F5041C5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F3ECBDF1-3AAD-418B-A005-E9435F5041C5}.Release|x64.ActiveCfg = Release|Any CPU
+		{F3ECBDF1-3AAD-418B-A005-E9435F5041C5}.Release|x64.Build.0 = Release|Any CPU
+		{F3ECBDF1-3AAD-418B-A005-E9435F5041C5}.Release|x86.ActiveCfg = Release|Any CPU
+		{F3ECBDF1-3AAD-418B-A005-E9435F5041C5}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -313,6 +341,8 @@ Global
 		{795D3F7B-F675-4612-928D-0640996D7149} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 		{91B56D43-36DD-4821-8D6E-0CD2461D1FAF} = {0AB3BF05-4346-4AA6-1389-037BE0695223}
 		{52C0B54B-9C1A-4D4F-B7F3-99CE3960AE95} = {0AB3BF05-4346-4AA6-1389-037BE0695223}
+		{27C83E37-2B9D-4B47-9096-A503C0D53F87} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
+		{F3ECBDF1-3AAD-418B-A005-E9435F5041C5} = {0AB3BF05-4346-4AA6-1389-037BE0695223}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {C753CED5-7508-4B2A-A107-C1E530FD786B}
