+++ src/Progesi.LiveDataExchange/Progesi.LiveDataExchange.csproj
@@ -1,0 +1,21 @@
+<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <TargetFramework>net48</TargetFramework>
+    <LangVersion>8.0</LangVersion>
+    <Nullable>disable</Nullable>
+    <RootNamespace>Progesi.LiveDataExchange</RootNamespace>
+    <AssemblyName>Progesi.LiveDataExchange</AssemblyName>
+    <IsPackable>false</IsPackable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\ProgesiCore\ProgesiCore.csproj" />
+    <ProjectReference Include="..\Progesi.GhExcelReadContract\Progesi.GhExcelReadContract.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="ClosedXML" Version="0.104.0" />
+    <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
+    <PackageReference Include="System.Data.SQLite.Core" Version="1.0.118.0" />
+  </ItemGroup>
+</Project>
