+++ src/ProgesiGrasshopperAssembly/Components/ProgesiDataExchangeComponent.cs
@@ -13,6 +13,7 @@ using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using ProgesiCore;
 using Progesi.GhExcelReadContract;
+using Progesi.LiveDataExchange;
 using ProgesiRepositories.Rhino;
 using ProgesiGrasshopperAssembly.Infrastructure; // ServiceHub, ProgesiIcons, MetadataRepositoryCompatExtensions
 using Rhino;
@@ -36,8 +37,7 @@ namespace ProgesiGrasshopperAssembly.Components
 {
   public sealed class ProgesiDataExchangeComponent : GH_Component
   {
-    // static ctor – eseguito una volta quando GH carica l'assembly
-  
+    private static readonly IGeometryValueCodec GeometryCodec = new RhinoGeometryValueCodecAdapter();
 
     public ProgesiDataExchangeComponent()
       : base("ProgesiData-Excel", "DataEx",
@@ -251,137 +251,7 @@ namespace ProgesiGrasshopperAssembly.Components
 
     private static (string path, string info) ExportExcel(string inPath, bool overwrite)
     {
-      var doc = RhinoDoc.ActiveDoc ?? throw new InvalidOperationException("RhinoDoc.ActiveDoc is null.");
-      StringTable table = doc.Strings ?? throw new InvalidOperationException("RhinoDoc.Strings is null.");
-
-      string p = NormalizeExportPath(inPath);
-      if (File.Exists(p))
-      {
-        if (!overwrite) throw new InvalidOperationException("File already exists: " + p);
-        try { using (var _ = File.Open(p, FileMode.Open, FileAccess.Write, FileShare.None)) { } } catch { }
-      }
-
-      var vars = ReadAllVarsFromTable(table);
-      var metas = ReadAllMetasFromTable(table);
-      var clusters = ReadAllClustersFromTable(table);
-      int unsupportedVarExports = vars.Count(v => v.IsExcelUnsupported);
-      var objectChunkRows = new List<GhExcelObjectSheet.ObjectChunkRow>();
-      foreach (var v in vars)
-      {
-        if (string.IsNullOrWhiteSpace(v.ObjectPayloadJson))
-          continue;
-        objectChunkRows.AddRange(GhExcelObjectSheet.ChunkPayload(v.Id, v.ObjectType, v.ObjectPayloadJson));
-      }
-      using (var wb = new XLWorkbook())
-      {
-        // Variables
-        var wsV = wb.Worksheets.Add("ProgesiVariables");
-        wsV.Cell(1, 1).Value = "Id";
-        wsV.Cell(1, 2).Value = "Hash";
-        wsV.Cell(1, 3).Value = "Name";
-        wsV.Cell(1, 4).Value = "Value";
-        wsV.Cell(1, 5).Value = "ValC";
-        wsV.Cell(1, 6).Value = "MetaId";
-        wsV.Cell(1, 7).Value = "Depends";
-        wsV.Cell(1, 8).Value = "Assumption";
-        int r = 2;
-        foreach (var v in vars)
-        {
-          wsV.Cell(r, 1).Value = v.Id;
-          wsV.Cell(r, 2).Value = v.Hash ?? "";
-          wsV.Cell(r, 3).Value = v.Name ?? "";
-          wsV.Cell(r, 4).Value = v.Value ?? "";
-          wsV.Cell(r, 5).Value = v.ValC ?? "";
-          wsV.Cell(r, 6).Value = FormatMetaIdForExcel(v.MetadataIds);
-          wsV.Cell(r, 7).Value = (v.Depends != null && v.Depends.Length > 0) ? string.Join(",", v.Depends) : "";
-          wsV.Cell(r, 8).Value = v.Assumption ? 1 : 0;
-          r++;
-        }
-        wsV.Columns().AdjustToContents();
-
-        // Metadata
-        var wsM = wb.Worksheets.Add("ProgesiMetadata");
-        wsM.Cell(1, 1).Value = "Id";
-        wsM.Cell(1, 2).Value = "Hash";
-        wsM.Cell(1, 3).Value = "By";
-        wsM.Cell(1, 4).Value = "Description";
-        wsM.Cell(1, 5).Value = "Refs";
-        wsM.Cell(1, 6).Value = "LM";
-        int r2 = 2;
-        foreach (var m in metas)
-        {
-          wsM.Cell(r2, 1).Value = m.Id;
-          wsM.Cell(r2, 2).Value = m.Hash ?? "";
-          wsM.Cell(r2, 3).Value = m.By ?? "";
-          wsM.Cell(r2, 4).Value = m.Description ?? "";
-          wsM.Cell(r2, 5).Value = (m.Refs != null && m.Refs.Length > 0) ? string.Join("|", m.Refs) : "";
-          wsM.Cell(r2, 6).Value = m.LM ?? "";
-          r2++;
-        }
-        wsM.Columns().AdjustToContents();
-
-        // Clusters
-        var wsC = wb.Worksheets.Add("ProgesiClusters");
-        wsC.Cell(1, 1).Value = "Id";
-        wsC.Cell(1, 2).Value = "Hash";
-        wsC.Cell(1, 3).Value = "Name";
-        wsC.Cell(1, 4).Value = "Description";
-        wsC.Cell(1, 5).Value = "VariableIds";
-        int r3 = 2;
-        foreach (var c in clusters)
-        {
-          wsC.Cell(r3, 1).Value = c.Id;
-          wsC.Cell(r3, 2).Value = c.Hash ?? "";
-          wsC.Cell(r3, 3).Value = c.Name ?? "";
-          wsC.Cell(r3, 4).Value = c.Description ?? "";
-          wsC.Cell(r3, 5).Value = (c.VariableIds != null && c.VariableIds.Length > 0)
-            ? string.Join(",", c.VariableIds)
-            : "";
-          r3++;
-        }
-        wsC.Columns().AdjustToContents();
-
-        if (objectChunkRows.Count > 0)
-        {
-          var wsO = wb.Worksheets.Add(GhExcelSheetNames.VariableObjects);
-          wsO.Cell(1, 1).Value = "VarId";
-          wsO.Cell(1, 2).Value = "ChunkIndex";
-          wsO.Cell(1, 3).Value = "ChunkCount";
-          wsO.Cell(1, 4).Value = "ObjectType";
-          wsO.Cell(1, 5).Value = "Payload";
-          int r4 = 2;
-          foreach (var chunk in objectChunkRows)
-          {
-            wsO.Cell(r4, 1).Value = chunk.VarId;
-            wsO.Cell(r4, 2).Value = chunk.ChunkIndex;
-            wsO.Cell(r4, 3).Value = chunk.ChunkCount;
-            wsO.Cell(r4, 4).Value = chunk.ObjectType ?? "";
-            wsO.Cell(r4, 5).Value = chunk.Payload ?? "";
-            r4++;
-          }
-          wsO.Columns().AdjustToContents();
-        }
-
-        wb.SaveAs(p);
-      }
-
-      string info = $"OK ExportExcel → {p} (Vars:{vars.Length}, Meta:{metas.Length}, Clusters:{clusters.Length}, ObjectChunks:{objectChunkRows.Count})";
-      if (unsupportedVarExports > 0)
-        info += $" | UnsupportedVarValues:{unsupportedVarExports}";
-      return (p, info);
-    }
-    private static string NormalizeExportPath(string inPath)
-    {
-      string p = (inPath ?? "").Trim();
-      if (string.IsNullOrEmpty(p))
-      {
-        string home = Environment.GetFolderPath(Environment.SpecialFolder.Desktop);
-        if (string.IsNullOrEmpty(home)) home = AppDomain.CurrentDomain.BaseDirectory;
-        p = Path.Combine(home, "Progesi_Export.xlsx");
-      }
-      if (Directory.Exists(p)) p = Path.Combine(p, "Progesi_Export.xlsx");
-      if (Path.GetExtension(p).Length == 0) p = p.TrimEnd(' ', '.') + ".xlsx";
-      return p;
+      return LiveExchangeExcelExporter.Export(ReadSnapshotFromTable(), inPath, overwrite);
     }
 
     // ------------------------------ IMPORT --------------------------------
@@ -394,378 +264,21 @@ namespace ProgesiGrasshopperAssembly.Components
                     string info)
     ImportExcelValidated(string inPath, bool strict, bool failOnError, int maxErrors, string mapJson, bool dryRun)
     {
-      // --- setup ---
-      string p = (inPath ?? "").Trim();
-      if (string.IsNullOrEmpty(p)) throw new ArgumentException("Path .xlsx not specified.");
-      if (!File.Exists(p)) throw new FileNotFoundException("File not found", p);
-
-      var warnTree = new GH_Structure<GH_String>();
-      var errTree = new GH_Structure<GH_String>();
-      var counts = new GH_Structure<GH_String>();
-      var errRC = new GH_Structure<GH_Integer>(); // {branch; i}: [row, col]
-
-      var logLines = new List<string>();
-      Action<string, string> LOG = (lvl, msg) => logLines.Add($"[{DateTime.Now:yyyy-MM-dd HH:mm:ss}] {lvl}: {msg}");
-      Action<int, string> WARN = (br, msg) => { warnTree.Append(new GH_String(msg), new GH_Path(br)); LOG("WARN", msg); };
-      Action<int, string> ERR = (br, msg) => { errTree.Append(new GH_String(msg), new GH_Path(br)); LOG("ERROR", msg); };
-      Action<int, int, int> AddErrRC = (br, row, col) =>
-      {
-        // un solo branch per sezione (0=Meta, 1=Vars), due interi consecutivi per ogni errore [row,col]
-        var path = new GH_Path(br);
-        errRC.Append(new GH_Integer(row), path);
-        errRC.Append(new GH_Integer(col), path);
-      };
-
-
       object repo; string hubInfo;
       if (!ServiceHub.TryGetMetadataRepository(out repo, out hubInfo))
         throw new InvalidOperationException("RHINO repository not available.");
 
-      // alias
-      var (varAliases, metaAliases) = GhExcelAliasMaps.Build(mapJson);
-
-      // limiti/regex
-      const int NAME_MAX = 128;
-      const int DESC_MAX = 512;
-      Func<string, bool> IsPrintable = s => string.IsNullOrEmpty(s) || s.All(ch => ch >= 32 && ch != 127);
-      Func<string, bool> IsHttpUrl = s => Uri.TryCreate(s, UriKind.Absolute, out var u) && (u.Scheme == "http" || u.Scheme == "https");
-      Func<string, bool> IsAbsPath = s => !string.IsNullOrEmpty(s) && Path.IsPathRooted(s);
-
-      // contatori
-      int metaRows = 0, metaOk = 0, metaWarn = 0, metaErr = 0, maxMetaId = 0;
-      int varRows = 0, varOk = 0, varWarn = 0, varErr = 0, maxVarId = 0;
-      int clusterRows = 0, clusterOk = 0, clusterWarn = 0, clusterErr = 0;
-
-      using (var wb = new XLWorkbook(p))
-      {
-        // ========== METADATA ==========
-        var wsM = GhExcelWorksheetLocator.TryGetWorksheet(wb, GhExcelSheetNames.Metadata, GhExcelSheetNames.MetadataAlias);
-        bool metaHeaderError = false;
-        Dictionary<string, int> mapM = null;
-        int r0M = 1, rNM = 0;
-
-        if (wsM == null)
-        {
-          string m = "Sheet 'ProgesiMetadata' not found.";
-          if (strict) { ERR(0, m); metaHeaderError = true; }
-          else { WARN(0, m); }
-        }
-        else
-        {
-          var headerM = GhExcelHeaderMap.Build(wsM, out r0M, out rNM);
-          mapM = GhExcelColumnMap.ResolveColumns(headerM, metaAliases);
-          var missingMeta = GhExcelColumnMap.MissingRequired(mapM, new[] { "BY", "DESCRIPTION" });
-          if (missingMeta.Count > 0)
-          {
-            string m = "Missing headers (Meta): " + string.Join(",", missingMeta);
-            if (strict) { ERR(0, m); metaHeaderError = true; AddErrRC(0, r0M, -1); }
-            else { WARN(0, m); metaWarn += missingMeta.Count; }
-          }
-        }
-
-        if (!metaHeaderError && wsM != null)
-        {
-          for (int r = r0M + 1; r <= rNM; r++)
-          {
-            string by = GhExcelCellReader.ReadCell(wsM, r, mapM, "BY");
-            string desc = GhExcelCellReader.ReadCell(wsM, r, mapM, "DESCRIPTION");
-            string refs = GhExcelCellReader.ReadCell(wsM, r, mapM, "REFS");
-            int id = GhExcelValueParsing.ToInt(GhExcelCellReader.ReadCell(wsM, r, mapM, "ID"));
-
-            if (GhExcelValueParsing.IsBlank(by) && GhExcelValueParsing.IsBlank(desc) && GhExcelValueParsing.IsBlank(refs))
-            { WARN(0, $"[Meta R{r}] empty row → skip"); metaWarn++; metaRows++; continue; }
-
-            // controlli extra
-            if (by.Length > NAME_MAX || !IsPrintable(by))
-            { var msg = $"[Meta R{r}] BY invalid (len/charset)"; (strict ? ERR : WARN)(0, msg); AddErrRC(0, r, mapM.TryGetValue("BY", out var c) ? c : 0); if (strict) { metaErr++; continue; } else { metaWarn++; } }
-
-            if (desc.Length > DESC_MAX || !IsPrintable(desc))
-            { var msg = $"[Meta R{r}] DESCRIPTION invalid (len/charset)"; (strict ? ERR : WARN)(0, msg); AddErrRC(0, r, mapM.TryGetValue("DESCRIPTION", out var c) ? c : 0); if (strict) { metaErr++; continue; } else { metaWarn++; } }
-
-            if (!string.IsNullOrWhiteSpace(refs))
-            {
-              foreach (var token in refs.Split('|'))
-              {
-                var s = token?.Trim(); if (string.IsNullOrEmpty(s)) continue;
-                if (!(IsHttpUrl(s) || IsAbsPath(s)))
-                { var msg = $"[Meta R{r}] REF invalid: {s}"; (strict ? ERR : WARN)(0, msg); AddErrRC(0, r, mapM.TryGetValue("REFS", out var c) ? c : 0); if (strict) { metaErr++; goto NEXT_META; } else { metaWarn++; } }
-              }
-            }
-
-            if (!dryRun)
-            {
-              var payload = new { id = id, by = by ?? "", info = desc ?? "", rf = refs ?? "", sn = "" };
-              object persisted; string upInfo;
-              bool ok = MetadataRepositoryCompatExtensions.TryUpsert(repo, payload, out persisted, out upInfo);
-              if (!ok)
-              { var msg = $"[Meta R{r}] import failed: {upInfo ?? "unknown"}"; ERR(0, msg); AddErrRC(0, r, 0); metaErr++; metaRows++; continue; }
-              int pid = 0; ReadPersistedId(persisted, ref pid);
-              if (pid > 0 && pid > maxMetaId) maxMetaId = pid;
-            }
-
-            metaOk++; metaRows++;
-            NEXT_META:;
-          }
-        }
-
-        // ========== VARIABLES ==========
-        var wsV = GhExcelWorksheetLocator.TryGetWorksheet(wb, GhExcelSheetNames.Variables, GhExcelSheetNames.VariablesAlias);
-        bool varHeaderError = false;
-        Dictionary<string, int> mapV = null;
-        int r0V = 1, rNV = 0;
-
-        if (wsV == null)
-        {
-          string m = "Sheet 'ProgesiVariables' not found.";
-          if (strict) { ERR(1, m); varHeaderError = true; }
-          else { WARN(1, m); }
-        }
-        else
-        {
-          var headerV = GhExcelHeaderMap.Build(wsV, out r0V, out rNV);
-          mapV = GhExcelColumnMap.ResolveColumns(headerV, varAliases);
-          var missingVar = GhExcelColumnMap.MissingRequired(mapV, new[] { "NAME", "VALUE" });
-          if (missingVar.Count > 0)
-          {
-            string m = "Missing headers (Vars): " + string.Join(",", missingVar);
-            if (strict) { ERR(1, m); varHeaderError = true; AddErrRC(1, r0V, -1); }
-            else { WARN(1, m); varWarn += missingVar.Count; }
-          }
-        }
-
-        if (!varHeaderError && wsV != null)
-        {
-          var objectChunkRows = ReadObjectChunkRows(wb);
-
-          for (int r = r0V + 1; r <= rNV; r++)
-          {
-            string name = GhExcelCellReader.ReadCell(wsV, r, mapV, "NAME");
-            string value = GhExcelCellReader.ReadCell(wsV, r, mapV, "VALUE");
-            string deps = GhExcelCellReader.ReadCell(wsV, r, mapV, "DEPENDS");
-            string asS = GhExcelCellReader.ReadCell(wsV, r, mapV, "ASSUMPTION");
-            int id = GhExcelValueParsing.ToInt(GhExcelCellReader.ReadCell(wsV, r, mapV, "ID"));
-            string metaCell = GhExcelCellReader.ReadCell(wsV, r, mapV, "METAID");
-            int[] metaIds = ParseMetadataIds(metaCell);
-
-            if (GhExcelValueParsing.IsBlank(name) && GhExcelValueParsing.IsBlank(value) && GhExcelValueParsing.IsBlank(deps) && GhExcelValueParsing.IsBlank(asS))
-            { WARN(1, $"[Var R{r}] empty row → skip"); varWarn++; varRows++; continue; }
-
-            // extra checks
-            if (name.Length > NAME_MAX || !IsPrintable(name))
-            { var msg = $"[Var R{r}] NAME invalid (len/charset)"; (strict ? ERR : WARN)(1, msg); AddErrRC(1, r, mapV.TryGetValue("NAME", out var c) ? c : 0); if (strict) { varErr++; continue; } else { varWarn++; } }
-
-            // MetaId existence (se fornito)
-            if (metaIds.Length > 0)
-            {
-              foreach (var mid in metaIds)
-              {
-                object dummy; string lookupInfo;
-                bool okMeta = MetadataRepositoryCompatExtensions.TryGetByHashThenId(repo, "", mid, out dummy, out lookupInfo);
-
-                if (!okMeta)
-                {
-                  var msg = $"[Var R{r}] METAID not found: {mid}";
-                  if (strict) { ERR(1, msg); AddErrRC(1, r, mapV.TryGetValue("METAID", out var c) ? c : 0); varErr++; metaIds = Array.Empty<int>(); break; }
-                  else { WARN(1, msg); AddErrRC(1, r, mapV.TryGetValue("METAID", out var c) ? c : 0); varWarn++; metaIds = Array.Empty<int>(); break; }
-                }
-              }
-              if (metaIds.Length == 0 && strict) { varRows++; continue; }
-            }
-
-            int[] depArr = GhExcelValueParsing.ParseDepends(deps);
-            bool ass = GhExcelValueParsing.ToBool(asS);
-
-            string geometryJson = null;
-            if (GhExcelObjectSheet.TryParseObjectMarker(value, out _))
-            {
-              if (!GhExcelObjectSheet.TryReassemblePayload(objectChunkRows, id, out var payload, out var objectType, out var objErr))
-              {
-                var msg = $"[Var R{r}] object payload missing/invalid: {objErr}";
-                if (strict) { ERR(1, msg); AddErrRC(1, r, mapV.TryGetValue("VALUE", out var c) ? c : 0); varErr++; }
-                else { WARN(1, msg); AddErrRC(1, r, mapV.TryGetValue("VALUE", out var c) ? c : 0); varWarn++; }
-                varRows++;
-                continue;
-              }
-
-              if (!ProgesiGeometryValueCodec.TryDecode(payload, out var geometry) || geometry == null)
-              {
-                var msg = $"[Var R{r}] object decode failed ({objectType})";
-                if (strict) { ERR(1, msg); AddErrRC(1, r, mapV.TryGetValue("VALUE", out var c) ? c : 0); varErr++; }
-                else { WARN(1, msg); AddErrRC(1, r, mapV.TryGetValue("VALUE", out var c) ? c : 0); varWarn++; }
-                varRows++;
-                continue;
-              }
-
-              geometryJson = payload;
-            }
-            else if (!GhExcelVariableValueSupport.IsImportSupported(value, out var unsupportedKind, out var unsupportedDetail))
-            {
-              var msg = GhExcelVariableValueSupport.BuildImportSkipMessage(r, unsupportedKind, unsupportedDetail);
-              if (strict) { ERR(1, msg); AddErrRC(1, r, mapV.TryGetValue("VALUE", out var c) ? c : 0); varErr++; }
-              else { WARN(1, msg); AddErrRC(1, r, mapV.TryGetValue("VALUE", out var c) ? c : 0); varWarn++; }
-              varRows++;
-              continue;
-            }
-
-            if (!dryRun)
-            {
-              var payload = new
-              {
-                id = id,
-                name = name ?? "",
-                value = geometryJson ?? (value ?? ""),
-                geometryJson = geometryJson ?? "",
-                unit = "",
-                by = "",
-                isAssumption = ass,
-                metadataIds = (object)metaIds,
-                mid = FormatMetaIdForExcel(metaIds),
-                depends = (object)depArr
-              };
-
-              object persisted; string upInfo;
-              bool ok = MetadataRepositoryCompatExtensions.TryUpsertVariable(repo, payload, out persisted, out upInfo);
-              if (!ok)
-              { var msg = $"[Var R{r}] import failed: {upInfo ?? "unknown"}"; ERR(1, msg); AddErrRC(1, r, 0); varErr++; varRows++; continue; }
-              int pid = 0; ReadPersistedId(persisted, ref pid);
-              if (pid > 0 && pid > maxVarId) maxVarId = pid;
-            }
-
-            varOk++; varRows++;
-          }
-        }
-
-        // ========== CLUSTERS ==========
-        var wsC = GhExcelWorksheetLocator.TryGetWorksheet(wb, GhExcelSheetNames.Clusters, GhExcelSheetNames.ClustersAlias);
-        bool clusterHeaderError = false;
-        Dictionary<string, int> mapC = null;
-        int r0C = 1, rNC = 0;
-
-        if (wsC == null)
-        {
-          string m = "Sheet 'ProgesiClusters' not found.";
-          if (strict) { ERR(2, m); clusterHeaderError = true; }
-          else { WARN(2, m); }
-        }
-        else
-        {
-          var headerC = GhExcelHeaderMap.Build(wsC, out r0C, out rNC);
-          mapC = GhExcelClusterSheet.ResolveClusterColumns(headerC);
-          var missingCluster = GhExcelColumnMap.MissingRequired(mapC, new[] { "ID" });
-          if (missingCluster.Count > 0)
-          {
-            string m = "Missing headers (Clusters): " + string.Join(",", missingCluster);
-            if (strict) { ERR(2, m); clusterHeaderError = true; AddErrRC(2, r0C, -1); }
-            else { WARN(2, m); clusterWarn += missingCluster.Count; }
-          }
-        }
-
-        StringTable clusterTable = null;
-        if (!dryRun && wsC != null && !clusterHeaderError)
-        {
-          var clusterDoc = RhinoDoc.ActiveDoc ?? throw new InvalidOperationException("RhinoDoc.ActiveDoc is null.");
-          clusterTable = clusterDoc.Strings ?? throw new InvalidOperationException("RhinoDoc.Strings is null.");
-        }
-
-        if (!clusterHeaderError && wsC != null)
-        {
-          for (int r = r0C + 1; r <= rNC; r++)
-          {
-            if (GhExcelClusterImport.IsBlankDataRow(wsC, r, mapC))
-            { WARN(2, $"[Cluster R{r}] empty row → skip"); clusterWarn++; clusterRows++; continue; }
-
-            if (!GhExcelClusterImport.TryBuildRowDto(wsC, r, mapC, out var rowDto, out var parseError))
-            {
-              var msg = $"[Cluster R{r}] {parseError}";
-              if (strict) { ERR(2, msg); AddErrRC(2, r, mapC.TryGetValue("ID", out var cidCol) ? cidCol : 0); clusterErr++; }
-              else { WARN(2, msg); clusterWarn++; }
-              clusterRows++;
-              continue;
-            }
-
-            if (!string.IsNullOrWhiteSpace(rowDto.ParseWarning))
-            {
-              var msg = $"[Cluster R{r}] {rowDto.ParseWarning}";
-              if (strict) { ERR(2, msg); AddErrRC(2, r, mapC.TryGetValue("VARIABLEIDS", out var vidCol) ? vidCol : 0); clusterErr++; clusterRows++; continue; }
-              else { WARN(2, msg); clusterWarn++; }
-            }
-
-            if (rowDto.VariableIds == null || rowDto.VariableIds.Length == 0)
-            {
-              WARN(2, $"[Cluster R{r}] no VariableIds → skip persist");
-              clusterWarn++;
-              clusterRows++;
-              continue;
-            }
-
-            string cName = rowDto.Name ?? "";
-            if (cName.Length > NAME_MAX || !IsPrintable(cName))
-            {
-              var msg = $"[Cluster R{r}] NAME invalid (len/charset)";
-              if (strict) { ERR(2, msg); AddErrRC(2, r, mapC.TryGetValue("NAME", out var nameCol) ? nameCol : 0); clusterErr++; clusterRows++; continue; }
-              else { WARN(2, msg); clusterWarn++; }
-            }
-
-            string cDesc = rowDto.Description ?? "";
-            if (cDesc.Length > DESC_MAX || !IsPrintable(cDesc))
-            {
-              var msg = $"[Cluster R{r}] DESCRIPTION invalid (len/charset)";
-              if (strict) { ERR(2, msg); AddErrRC(2, r, mapC.TryGetValue("DESCRIPTION", out var descCol) ? descCol : 0); clusterErr++; clusterRows++; continue; }
-              else { WARN(2, msg); clusterWarn++; }
-            }
-
-            if (!dryRun && clusterTable != null)
-            {
-              var clusterDto = new ClusterDto
-              {
-                Id = rowDto.Id,
-                Name = cName,
-                Description = cDesc,
-                VariableIds = rowDto.VariableIds,
-                Hashtag = rowDto.Hashtag
-              };
-
-              string json = JsonConvert.SerializeObject(clusterDto);
-              clusterTable.SetString("Progesi.Cluster", "cluster:" + rowDto.Id.ToString(CultureInfo.InvariantCulture), json);
-
-              int next = ReadCounter(clusterTable, "Progesi.Cluster");
-              if (rowDto.Id + 1 > next)
-                clusterTable.SetString("Progesi.Cluster", "__next__", (rowDto.Id + 1).ToString(CultureInfo.InvariantCulture));
-            }
-
-            clusterOk++;
-            clusterRows++;
-          }
-        }
-      } // using wb
-
-      // Aggiorna contatori solo se NON è DryRun
+      StringTable clusterTable = null;
       if (!dryRun)
       {
-        var doc = RhinoDoc.ActiveDoc ?? throw new InvalidOperationException("RhinoDoc.ActiveDoc is null.");
-        StringTable table = doc.Strings ?? throw new InvalidOperationException("RhinoDoc.Strings is null.");
-        int curMetaNext = ReadCounter(table, "Progesi.Meta");
-        int curVarNext = ReadCounter(table, "Progesi.Var");
-        if (maxMetaId > 0) table.SetString("Progesi.Meta", "__next__", Math.Max(curMetaNext, maxMetaId + 1).ToString(CultureInfo.InvariantCulture));
-        if (maxVarId > 0) table.SetString("Progesi.Var", "__next__", Math.Max(curVarNext, maxVarId + 1).ToString(CultureInfo.InvariantCulture));
+        var clusterDoc = RhinoDoc.ActiveDoc ?? throw new InvalidOperationException("RhinoDoc.ActiveDoc is null.");
+        clusterTable = clusterDoc.Strings ?? throw new InvalidOperationException("RhinoDoc.Strings is null.");
       }
 
-      // log
-      string logPath = p + ".import.log.txt";
-      try { File.WriteAllLines(logPath, logLines, Encoding.UTF8); } catch { logPath = ""; }
-
-      // counts
-      counts.Append(new GH_String($"Meta rows={metaRows} ok={metaOk} warn={metaWarn} err={metaErr}"), new GH_Path(0));
-      counts.Append(new GH_String($"Vars rows={varRows} ok={varOk} warn={varWarn} err={varErr}"), new GH_Path(1));
-      counts.Append(new GH_String($"Clusters rows={clusterRows} ok={clusterOk} warn={clusterWarn} err={clusterErr}"), new GH_Path(2));
-
-      string prefix = dryRun ? "PREVIEW " : "OK ";
-      string info = $"{prefix}ImportExcel ← {p} | Meta {metaOk}/{metaRows} (warn:{metaWarn}, err:{metaErr}) | " +
-                    $"Vars {varOk}/{varRows} (warn:{varWarn}, err:{varErr}) | " +
-                    $"Clusters {clusterOk}/{clusterRows} (warn:{clusterWarn}, err:{clusterErr}) | " +
-                    $"Log: {(string.IsNullOrWhiteSpace(logPath) ? "-" : logPath)}";
-
-      return (p, logPath, warnTree, errTree, counts, errRC, info);
+      var sink = new RhinoLiveExchangeImportSink(repo, clusterTable);
+      var result = LiveExchangeExcelImporter.ImportValidated(
+        inPath, strict, failOnError, maxErrors, mapJson, dryRun, sink, GeometryCodec);
+      return MapImportResult(result);
     }
 
     // =========================== SQLITE – EXPORT ===========================
@@ -813,251 +326,9 @@ namespace ProgesiGrasshopperAssembly.Components
 
     // in cima al file, tra gli using:
 
-    // ...dentro la classe ProgesiDataExchangeComponent sostituisci l'intero metodo ExportSqlite con questo:
     private static (string path, string info) ExportSqlite(string inPath, bool overwrite)
     {
-      // 0) Sorgente: RHINO StringTable
-      var doc = RhinoDoc.ActiveDoc ?? throw new InvalidOperationException("RhinoDoc.ActiveDoc is null.");
-      var table = doc.Strings ?? throw new InvalidOperationException("RhinoDoc.Strings is null.");
-
-      // 1) Leggi tutto subito (una sola passata)
-      var metas = ReadAllMetasFromTable(table);   // MetaRow[]
-      var vars = ReadAllVarsFromTable(table);    // VarRow[]
-      var metaIdsPresent = new HashSet<int>(metas.Select(m => m.Id));
-      var clusters = ReadAllClustersFromTable(table);
-
-      // 2) Prepara percorso destinazione
-      string p = (inPath ?? string.Empty).Trim();
-      if (string.IsNullOrWhiteSpace(p))                       // <-- fix qui
-      {
-        var desktop = Environment.GetFolderPath(Environment.SpecialFolder.Desktop);
-        if (string.IsNullOrEmpty(desktop)) desktop = AppDomain.CurrentDomain.BaseDirectory;
-        p = Path.Combine(desktop, "Progesi_Export.db");
-      }
-      if (Directory.Exists(p))
-        p = Path.Combine(p, "Progesi_Export.db");
-
-      if (File.Exists(p))
-      {
-        if (!overwrite)
-          throw new InvalidOperationException("Il file esiste già: " + p);
-
-        // prova a rimuoverlo; se è lockato, genera un nome alternativo con timestamp
-        try
-        {
-          using (var fs = File.Open(p, FileMode.Open, FileAccess.ReadWrite, FileShare.None)) { }
-          File.Delete(p);
-        }
-        catch
-        {
-          var dir = Path.GetDirectoryName(p) ?? ".";
-          var name = Path.GetFileNameWithoutExtension(p);
-          var ext = Path.GetExtension(p);
-          var stamp = DateTime.Now.ToString("yyyyMMdd_HHmmss");
-          p = Path.Combine(dir, $"{name}_{stamp}{ext}");
-        }
-      }
-
-      // 3) Scrittura in un'unica transazione
-      using (var cn = new SQLiteConnection($@"Data Source={p};Version=3;"))
-      {
-        cn.Open();
-
-        using (var tx = cn.BeginTransaction())
-        using (var cmd = new SQLiteCommand(cn))
-        {
-          // abilita FK e crea schema senza UNIQUE su Hash
-          cmd.CommandText = "PRAGMA foreign_keys=ON;";
-          cmd.ExecuteNonQuery();
-
-          cmd.CommandText = @"
-CREATE TABLE IF NOT EXISTS Metadata (
-  Id           INTEGER PRIMARY KEY,
-  Hash         TEXT NOT NULL,
-  By           TEXT,
-  Description  TEXT,
-  LM           TEXT
-);
-CREATE TABLE IF NOT EXISTS Variables (
-  Id           INTEGER PRIMARY KEY,
-  Hash         TEXT NOT NULL,
-  Name         TEXT NOT NULL,
-  Value        TEXT,
-  ValC         TEXT,
-  MetaId       INTEGER NULL,
-  Assumption   INTEGER NOT NULL DEFAULT 0,
-  FOREIGN KEY (MetaId) REFERENCES Metadata(Id) ON DELETE SET NULL
-);
-CREATE TABLE IF NOT EXISTS Refs (
-  MetaId       INTEGER NOT NULL,
-  Ref          TEXT NOT NULL,
-  PRIMARY KEY (MetaId, Ref),
-  FOREIGN KEY (MetaId) REFERENCES Metadata(Id) ON DELETE CASCADE
-);
-CREATE TABLE IF NOT EXISTS VariableDepends (
-  VarId        INTEGER NOT NULL,
-  DepId        INTEGER NOT NULL,
-  PRIMARY KEY (VarId, DepId),
-  FOREIGN KEY (VarId) REFERENCES Variables(Id) ON DELETE CASCADE
-);
-CREATE TABLE IF NOT EXISTS Clusters (
-  Id          INTEGER PRIMARY KEY,
-  Hash        TEXT NOT NULL,
-  Name        TEXT NOT NULL,
-  Description TEXT
-);
-
-CREATE TABLE IF NOT EXISTS ClusterVariables (
-  ClusterId   INTEGER NOT NULL,
-  VarId       INTEGER NOT NULL,
-  PRIMARY KEY (ClusterId, VarId),
-  FOREIGN KEY (ClusterId) REFERENCES Clusters(Id) ON DELETE CASCADE,
-  FOREIGN KEY (VarId)     REFERENCES Variables(Id) ON DELETE CASCADE
-);";
-          cmd.ExecuteNonQuery();
-
-          // 3.1) Metadata
-          cmd.CommandText = "INSERT OR REPLACE INTO Metadata (Id,Hash,By,Description,LM) VALUES (@id,@hash,@by,@descr,@lm)";
-          var pId = new SQLiteParameter("@id");
-          var pHash = new SQLiteParameter("@hash");
-          var pBy = new SQLiteParameter("@by");
-          var pDescr = new SQLiteParameter("@descr");
-          var pLM = new SQLiteParameter("@lm");
-          cmd.Parameters.AddRange(new[] { pId, pHash, pBy, pDescr, pLM });
-
-          foreach (var m in metas)
-          {
-            pId.Value = m.Id;
-            pHash.Value = m.Hash ?? string.Empty;
-            pBy.Value = m.By ?? string.Empty;
-            pDescr.Value = m.Description ?? string.Empty;
-            pLM.Value = m.LM ?? string.Empty;
-            cmd.ExecuteNonQuery();
-          }
-
-          // 3.2) Variabili (MetaId → NULL se non presente tra i metadata)
-          cmd.Parameters.Clear();
-          cmd.CommandText = "INSERT OR REPLACE INTO Variables (Id,Hash,Name,Value,ValC,MetaId,Assumption) VALUES (@id,@hash,@name,@value,@valc,@mid,@ass)";
-          var vId = new SQLiteParameter("@id");
-          var vHash = new SQLiteParameter("@hash");
-          var vName = new SQLiteParameter("@name");
-          var vVal = new SQLiteParameter("@value");
-          var vValC = new SQLiteParameter("@valc");
-          var vMid = new SQLiteParameter("@mid");
-          var vAss = new SQLiteParameter("@ass");
-          cmd.Parameters.AddRange(new[] { vId, vHash, vName, vVal, vValC, vMid, vAss });
-
-          foreach (var v in vars)
-          {
-            vId.Value = v.Id;
-            vHash.Value = v.Hash ?? string.Empty;
-            vName.Value = v.Name ?? string.Empty;
-            vVal.Value = v.Value ?? string.Empty;
-            vValC.Value = v.ValC ?? string.Empty;
-            vMid.Value = (v.MetaId > 0 && metaIdsPresent.Contains(v.MetaId)) ? (object)v.MetaId : (object)DBNull.Value;
-            vAss.Value = v.Assumption ? 1 : 0;
-            cmd.ExecuteNonQuery();
-          }
-
-          // 3.3) Refs
-          cmd.Parameters.Clear();
-          cmd.CommandText = "INSERT OR REPLACE INTO Refs (MetaId,Ref) VALUES (@mid,@ref)";
-          var rMid = new SQLiteParameter("@mid");
-          var rRef = new SQLiteParameter("@ref");
-          cmd.Parameters.AddRange(new[] { rMid, rRef });
-
-          foreach (var m in metas)
-          {
-            if (m.Refs == null || m.Refs.Length == 0) continue;
-            if (!metaIdsPresent.Contains(m.Id)) continue;
-
-            foreach (var rf in m.Refs)
-            {
-              rMid.Value = m.Id;
-              rRef.Value = rf ?? string.Empty;
-              cmd.ExecuteNonQuery();
-            }
-          }
-
-          // 3.4) VariableDepends
-          cmd.Parameters.Clear();
-          cmd.CommandText = "INSERT OR REPLACE INTO VariableDepends (VarId,DepId) VALUES (@vid,@did)";
-          var dVid = new SQLiteParameter("@vid");
-          var dDid = new SQLiteParameter("@did");
-          cmd.Parameters.AddRange(new[] { dVid, dDid });
-
-          var varIds = new HashSet<int>(vars.Select(v => v.Id));
-          foreach (var v in vars)
-          {
-            if (v.Depends == null || v.Depends.Length == 0) continue;
-            foreach (var dep in v.Depends)
-            {
-              if (!varIds.Contains(dep)) continue;
-              dVid.Value = v.Id;
-              dDid.Value = dep;
-              cmd.ExecuteNonQuery();
-            }
-          }
-
-          // 3.5) Clusters + reset join
-          cmd.Parameters.Clear();
-
-          // reset per export deterministico
-          cmd.CommandText = "DELETE FROM ClusterVariables;";
-          cmd.ExecuteNonQuery();
-
-          cmd.CommandText = "DELETE FROM Clusters;";
-          cmd.ExecuteNonQuery();
-
-          // insert clusters
-          cmd.CommandText = "INSERT OR REPLACE INTO Clusters (Id,Hash,Name,Description) VALUES (@id,@hash,@name,@desc)";
-          var cId = new SQLiteParameter("@id");
-          var cHash = new SQLiteParameter("@hash");
-          var cName = new SQLiteParameter("@name");
-          var cDesc = new SQLiteParameter("@desc");
-          cmd.Parameters.AddRange(new[] { cId, cHash, cName, cDesc });
-
-          foreach (var c in clusters)
-          {
-            cId.Value = c.Id;
-            cHash.Value = c.Hash ?? string.Empty;
-            cName.Value = c.Name ?? string.Empty;
-            cDesc.Value = c.Description ?? string.Empty;
-            cmd.ExecuteNonQuery();
-          }
-
-          // 3.6) ClusterVariables (join)
-          cmd.Parameters.Clear();
-          cmd.CommandText = "INSERT OR REPLACE INTO ClusterVariables (ClusterId,VarId) VALUES (@cid,@vid)";
-          var jCid = new SQLiteParameter("@cid");
-          var jVid = new SQLiteParameter("@vid");
-          cmd.Parameters.AddRange(new[] { jCid, jVid });
-
-          // usiamo solo VarId esistenti (coerente con depends)
-          foreach (var c in clusters)
-          {
-            if (c.VariableIds == null || c.VariableIds.Length == 0) continue;
-
-            foreach (var vid in c.VariableIds)
-            {
-              if (!varIds.Contains(vid)) continue; // varIds è HashSet<int> già creato sopra
-
-              jCid.Value = c.Id;
-              jVid.Value = vid;
-              cmd.ExecuteNonQuery();
-            }
-          }
-
-          tx.Commit();
-        }
-      }
-
-      // 4) Riepilogo
-      var info = string.Format(                                  // <-- fix Count/Length e firma di Format
-          CultureInfo.InvariantCulture,
-          "OK ExportSqlite → {0} (Meta:{1}, Vars:{2}, Clusters:{3})",
-          p, metas.Length, vars.Length, clusters.Length);
-      return (p, info);
+      return LiveExchangeSqliteExporter.Export(ReadSnapshotFromTable(), inPath, overwrite);
     }
 
     // =========================== SQLITE – IMPORT ===========================
@@ -1068,285 +339,64 @@ CREATE TABLE IF NOT EXISTS ClusterVariables (
                  GH_Structure<GH_String> countsOut,
                  GH_Structure<GH_Integer> errRcOut,
                  string info)
- ImportSqlite(string inDbPath, bool strict, bool dryRun)
+    ImportSqlite(string inDbPath, bool strict, bool dryRun)
     {
-      string db = (inDbPath ?? "").Trim();
-      if (string.IsNullOrEmpty(db)) throw new ArgumentException("SQLite path not specified.");
-      if (!File.Exists(db)) throw new FileNotFoundException("SQLite file not found.", db);
-
-      var warnTree = new GH_Structure<GH_String>();
-      var errTree = new GH_Structure<GH_String>();
-      var counts = new GH_Structure<GH_String>();
-      var errRC = new GH_Structure<GH_Integer>(); // {branch} → coppie [row,col]
-
-      var logLines = new List<string>();
-      Action<string, string> LOG = (lvl, msg) => logLines.Add($"[{DateTime.Now:yyyy-MM-dd HH:mm:ss}] {lvl}: {msg}");
-      Action<int, string> WARN = (br, msg) => { warnTree.Append(new GH_String(msg), new GH_Path(br)); LOG("WARN", msg); };
-      Action<int, string> ERR = (br, msg) => { errTree.Append(new GH_String(msg), new GH_Path(br)); LOG("ERROR", msg); };
-      Action<int, int, int> AddErrRC = (br, row, col) =>
-      {
-        var path = new GH_Path(br);
-        errRC.Append(new GH_Integer(row), path);
-        errRC.Append(new GH_Integer(col), path);
-      };
-
       object repo; string hub;
       if (!ServiceHub.TryGetMetadataRepository(out repo, out hub))
         throw new InvalidOperationException("RHINO repository not available.");
 
-      // controlli extra coerenti con Excel
-      const int NAME_MAX = 128;
-      const int DESC_MAX = 512;
-      Func<string, bool> IsPrintable = s => string.IsNullOrEmpty(s) || s.All(ch => ch >= 32 && ch != 127);
-      Func<string, bool> IsHttpUrl = s => Uri.TryCreate(s, UriKind.Absolute, out var u) && (u.Scheme == "http" || u.Scheme == "https");
-      Func<string, bool> IsAbsPath = s => !string.IsNullOrEmpty(s) && Path.IsPathRooted(s);
-
-      int metaRows = 0, metaOk = 0, metaWarn = 0, metaErr = 0;
-      int varRows = 0, varOk = 0, varWarn = 0, varErr = 0;
-      int clusterRows = 0, clusterOk = 0, clusterWarn = 0, clusterErr = 0;
-
-      using (var cn = new Microsoft.Data.Sqlite.SqliteConnection($"Data Source={db};Mode=ReadOnly"))
+      StringTable clusterTable = null;
+      if (!dryRun)
       {
-        cn.Open();
-
-        // Verifica schema
-        Func<string, bool> HasTable = name =>
-        {
-          using (var chk = cn.CreateCommand())
-          {
-            chk.CommandText = "SELECT name FROM sqlite_master WHERE type='table' AND name=@n";
-            chk.Parameters.AddWithValue("@n", name);
-            var r = chk.ExecuteScalar();
-            return r != null && r != DBNull.Value;
-          }
-        };
-
-        if (!HasTable("Metadata") || !HasTable("Variables"))
-        {
-          ERR(0, "SQLite schema not found (Metadata/Variables). Did export fail previously?");
-          string bad = db + ".import.log.txt";
-          try { File.WriteAllLines(bad, logLines, Encoding.UTF8); } catch { bad = ""; }
-          return (db, bad, warnTree, errTree, counts, errRC, "Error: missing tables in SQLite database.");
-        }
-
-        // ===== METADATA =====
-        using (var cmd = cn.CreateCommand())
-        {
-          cmd.CommandText = "SELECT Id, By, Description, LM FROM Metadata ORDER BY Id";
-          using (var rd = cmd.ExecuteReader())
-          {
-            int row = 0; // 1-based “visuale”
-            while (rd.Read())
-            {
-              row++; metaRows++;
-              int id = rd.IsDBNull(0) ? 0 : rd.GetInt32(0);
-              string by = rd.IsDBNull(1) ? "" : rd.GetString(1);
-              string ds = rd.IsDBNull(2) ? "" : rd.GetString(2);
-
-              // controlli extra
-              if (by.Length > NAME_MAX || !IsPrintable(by))
-              { var msg = $"[Meta R{row}] BY invalid (len/charset)"; (strict ? ERR : WARN)(0, msg); AddErrRC(0, row, 2); if (strict) { metaErr++; continue; } else { metaWarn++; } }
-
-              if (ds.Length > DESC_MAX || !IsPrintable(ds))
-              { var msg = $"[Meta R{row}] DESCRIPTION invalid (len/charset)"; (strict ? ERR : WARN)(0, msg); AddErrRC(0, row, 3); if (strict) { metaErr++; continue; } else { metaWarn++; } }
-
-              // refs: leggi e valida
-              string refsJoined = "";
-              using (var cmdR = cn.CreateCommand())
-              {
-                cmdR.CommandText = "SELECT Ref FROM Refs WHERE MetaId=@id";
-                cmdR.Parameters.AddWithValue("@id", id);
-                using (var rr = cmdR.ExecuteReader())
-                {
-                  var refs = new List<string>();
-                  while (rr.Read())
-                  {
-                    string rfs = rr.IsDBNull(0) ? "" : rr.GetString(0);
-                    if (!string.IsNullOrWhiteSpace(rfs))
-                    {
-                      if (!(IsHttpUrl(rfs) || IsAbsPath(rfs)))
-                      { var msg = $"[Meta R{row}] REF invalid: {rfs}"; (strict ? ERR : WARN)(0, msg); AddErrRC(0, row, 5); if (strict) { metaErr++; refs.Clear(); break; } else { metaWarn++; } }
-                      refs.Add(rfs);
-                    }
-                  }
-                  refsJoined = refs.Count > 0 ? string.Join("|", refs) : "";
-                }
-              }
-
-              if (!dryRun)
-              {
-                var payload = new { id = id, by = by ?? "", info = ds ?? "", rf = refsJoined, sn = "" };
-                object persisted; string upInfo;
-                bool ok = MetadataRepositoryCompatExtensions.TryUpsert(repo, payload, out persisted, out upInfo);
-                if (!ok) { ERR(0, $"[Meta R{row}] import failed: {upInfo ?? "unknown"}"); metaErr++; continue; }
-              }
-
-              metaOk++;
-            }
-          }
-        }
-
-        bool hasClusters = HasTable("Clusters") && HasTable("ClusterVariables");
-        // ===== VARIABLES =====
-        using (var cmd = cn.CreateCommand())
-        {
-          cmd.CommandText = "SELECT Id, Name, Value, MetaId, Assumption FROM Variables ORDER BY Id";
-          using (var rd = cmd.ExecuteReader())
-          {
-            int row = 0;
-            while (rd.Read())
-            {
-              row++; varRows++;
-              int id = rd.IsDBNull(0) ? 0 : rd.GetInt32(0);
-              string nm = rd.IsDBNull(1) ? "" : rd.GetString(1);
-              string vl = rd.IsDBNull(2) ? "" : rd.GetString(2);
-              int mid = rd.IsDBNull(3) ? 0 : rd.GetInt32(3);
-              bool ass = !rd.IsDBNull(4) && (rd.GetInt32(4) != 0);
-
-              // extra: NAME
-              if (nm.Length > NAME_MAX || !IsPrintable(nm))
-              { var msg = $"[Var R{row}] NAME invalid (len/charset)"; (strict ? ERR : WARN)(1, msg); AddErrRC(1, row, 2); if (strict) { varErr++; continue; } else { varWarn++; } }
-
-              // depends
-              int[] dep = Array.Empty<int>();
-              using (var cmdD = cn.CreateCommand())
-              {
-                cmdD.CommandText = "SELECT DepId FROM VariableDepends WHERE VarId=@id ORDER BY DepId";
-                cmdD.Parameters.AddWithValue("@id", id);
-                using (var rr = cmdD.ExecuteReader())
-                {
-                  var tmp = new List<int>(); while (rr.Read()) tmp.Add(rr.GetInt32(0)); dep = tmp.ToArray();
-                }
-              }
-
-              // metaId: se specificato deve esistere (preview = controllo; run = controllo + eventuale dedupe lato compat)
-              if (mid > 0)
-              {
-                object dummy; string lookupInfo;
-                bool okMeta = MetadataRepositoryCompatExtensions.TryGetByHashThenId(repo, "", mid, out dummy, out lookupInfo);
-                if (!okMeta)
-                { var msg = $"[Var R{row}] METAID not found: {mid}"; (strict ? ERR : WARN)(1, msg); AddErrRC(1, row, 4); if (strict) { varErr++; continue; } else { varWarn++; mid = 0; } }
-              }
-
-              if (!dryRun)
-              {
-                var payload = new
-                {
-                  id = id,
-                  name = nm ?? "",
-                  value = vl ?? "",
-                  unit = "",
-                  by = "",
-                  isAssumption = ass,
-                  mid = (mid > 0 ? mid.ToString(CultureInfo.InvariantCulture) : ""),
-                  depends = (object)dep
-                };
-                object persisted; string upInfo;
-                bool ok = MetadataRepositoryCompatExtensions.TryUpsertVariable(repo, payload, out persisted, out upInfo);
-                if (!ok) { ERR(1, $"[Var R{row}] import failed: {upInfo ?? "unknown"}"); varErr++; continue; }
-              }
-
-              varOk++;
-            }
-          }
-        }
-        if (hasClusters && !dryRun)
-        {
-          // Scriviamo direttamente nella StringTable Rhino come fa DataEx per Excel
-          var doc = RhinoDoc.ActiveDoc ?? throw new InvalidOperationException("RhinoDoc.ActiveDoc is null.");
-          var table = doc.Strings ?? throw new InvalidOperationException("RhinoDoc.Strings is null.");
-
-          using (var cmdC = cn.CreateCommand())
-          {
-            cmdC.CommandText = "SELECT Id, Name, Description, Hash FROM Clusters ORDER BY Id";
-            using (var rdC = cmdC.ExecuteReader())
-            {
-              while (rdC.Read())
-              {
-                clusterRows++;
-
-                int cid = rdC.IsDBNull(0) ? 0 : rdC.GetInt32(0);
-                string cnm = rdC.IsDBNull(1) ? "" : rdC.GetString(1);
-                string cds = rdC.IsDBNull(2) ? "" : rdC.GetString(2);
-                string ch = rdC.IsDBNull(3) ? "" : rdC.GetString(3);
-
-                // varIds join
-                var varIdsList = new List<int>();
-                using (var cmdJ = cn.CreateCommand())
-                {
-                  cmdJ.CommandText = "SELECT VarId FROM ClusterVariables WHERE ClusterId=@id ORDER BY VarId";
-                  cmdJ.Parameters.AddWithValue("@id", cid);
-                  using (var rdJ = cmdJ.ExecuteReader())
-                  {
-                    while (rdJ.Read())
-                    {
-                      int vid = rdJ.IsDBNull(0) ? 0 : rdJ.GetInt32(0);
-                      if (vid > 0) varIdsList.Add(vid);
-                    }
-                  }
-                }
-
-                var varIdsArr = varIdsList.Distinct().OrderBy(x => x).ToArray();
-
-                // ricalcolo hash coerente dal dominio (salviamo comunque ch in dto per traccia)
-                var cluster = ProgesiVariableCluster.Rehydrate(cid, cnm ?? "", varIdsArr, cds, string.IsNullOrWhiteSpace(ch) ? null : ch);
-
-                var dto = new ClusterDto
-                {
-                  Id = cid,
-                  Name = cnm ?? "",
-                  Description = cds ?? "",
-                  VariableIds = varIdsArr,
-                  Hashtag = cluster.Hashtag
-                };
-
-                string json = JsonConvert.SerializeObject(dto);
-
-                table.SetString("Progesi.Cluster", "cluster:" + cid.ToString(CultureInfo.InvariantCulture), json);
-
-                // bump __next__
-                int next = ReadCounter(table, "Progesi.Cluster");
-                if (cid + 1 > next)
-                  table.SetString("Progesi.Cluster", "__next__", (cid + 1).ToString(CultureInfo.InvariantCulture));
-
-                clusterOk++;
-              }
-            }
-          }
-
-          LOG("INFO", $"[Clusters] Imported {clusterOk}/{clusterRows} from SQLite.");
-          counts.Append(new GH_String($"Clusters rows={clusterRows} ok={clusterOk} warn={clusterWarn} err={clusterErr}"), new GH_Path(2));
-        }
-
+        var clusterDoc = RhinoDoc.ActiveDoc ?? throw new InvalidOperationException("RhinoDoc.ActiveDoc is null.");
+        clusterTable = clusterDoc.Strings ?? throw new InvalidOperationException("RhinoDoc.Strings is null.");
       }
 
-      // counts + log
-      counts.Append(new GH_String($"Meta rows={metaRows} ok={metaOk} warn={metaWarn} err={metaErr}"), new GH_Path(0));
-      counts.Append(new GH_String($"Vars rows={varRows} ok={varOk} warn={varWarn} err={varErr}"), new GH_Path(1));
-
-      string logPath = db + ".import.log.txt";
-      try { File.WriteAllLines(logPath, logLines, Encoding.UTF8); } catch { logPath = ""; }
-
-      string prefix = dryRun ? "PREVIEW " : "OK ";
-      string info = $"{prefix}ImportSqlite ← {db} | " +
-              $"Meta {metaOk}/{metaRows} (warn:{metaWarn}, err:{metaErr}) | " +
-              $"Vars {varOk}/{varRows} (warn:{varWarn}, err:{varErr}) | " +
-              $"Clusters {clusterOk}/{clusterRows} (warn:{clusterWarn}, err:{clusterErr}) | " +
-              $"Log: {(string.IsNullOrWhiteSpace(logPath) ? "-" : logPath)}";
-      return (db, logPath, warnTree, errTree, counts, errRC, info);
+      var sink = new RhinoLiveExchangeImportSink(repo, clusterTable);
+      var result = LiveExchangeSqliteImporter.ImportValidated(inDbPath, strict, dryRun, sink);
+      return MapImportResult(result);
     }
 
-    private static string ResolveSqlitePath(string inputPath)
+    private static LiveExchangeSnapshot ReadSnapshotFromTable()
     {
-      // Se è una directory, metti dentro un default filename
-      if (Directory.Exists(inputPath))
-        return Path.Combine(inputPath, "progesi.db");
+      var doc = RhinoDoc.ActiveDoc ?? throw new InvalidOperationException("RhinoDoc.ActiveDoc is null.");
+      StringTable table = doc.Strings ?? throw new InvalidOperationException("RhinoDoc.Strings is null.");
+      return new LiveExchangeSnapshot
+      {
+        Variables = ReadAllVarsFromTable(table),
+        Metadata = ReadAllMetasFromTable(table),
+        Clusters = ReadAllClustersFromTable(table)
+      };
+    }
 
-      // Se non ha estensione, aggiungi .db
-      if (string.IsNullOrWhiteSpace(Path.GetExtension(inputPath)))
-        return inputPath + ".db";
+    private static (string srcPath,
+                    string logPath,
+                    GH_Structure<GH_String> warnOut,
+                    GH_Structure<GH_String> errOut,
+                    GH_Structure<GH_String> countsOut,
+                    GH_Structure<GH_Integer> errRcOut,
+                    string info)
+    MapImportResult(LiveExchangeImportResult result)
+    {
+      var warnTree = new GH_Structure<GH_String>();
+      var errTree = new GH_Structure<GH_String>();
+      var counts = new GH_Structure<GH_String>();
+      var errRC = new GH_Structure<GH_Integer>();
 
-      return inputPath;
+      foreach (var w in result.Warnings)
+        warnTree.Append(new GH_String(w.Text), new GH_Path(w.Branch));
+      foreach (var e in result.Errors)
+        errTree.Append(new GH_String(e.Text), new GH_Path(e.Branch));
+      foreach (var c in result.Counts)
+        counts.Append(new GH_String(c.Text), new GH_Path(c.Branch));
+      foreach (var rc in result.ErrorCoordinates)
+      {
+        var path = new GH_Path(rc.Branch);
+        errRC.Append(new GH_Integer(rc.Row), path);
+        errRC.Append(new GH_Integer(rc.Col), path);
+      }
+
+      return (result.SourcePath, result.LogPath, warnTree, errTree, counts, errRC, result.Info);
     }
 
     private static void ReadPersistedId(object persisted, ref int target)
@@ -1389,31 +439,6 @@ CREATE TABLE IF NOT EXISTS ClusterVariables (
       public object[] Snips { get; set; }
     }
 
-    private sealed class VarRow
-    {
-      public int Id;
-      public string Hash;
-      public string Name;
-      public string Value;
-      public string ValC;
-      public int MetaId;
-      public int[] MetadataIds = Array.Empty<int>();
-      public int[] Depends;
-      public bool Assumption;
-      public bool IsExcelUnsupported;
-      public string ObjectType = "";
-      public string ObjectPayloadJson = "";
-    }
-
-    private sealed class MetaRow
-    {
-      public int Id;
-      public string Hash;
-      public string By;
-      public string Description;
-      public string[] Refs;
-      public string LM;
-    }
 #nullable enable
     private sealed class ClusterDto
     {
@@ -1423,20 +448,11 @@ CREATE TABLE IF NOT EXISTS ClusterVariables (
       public int[]? VariableIds { get; set; }
       public string? Hashtag { get; set; }
     }
-
-    private sealed class ClusterRow
-    {
-      public int Id;
-      public string Hash = "";
-      public string Name = "";
-      public string Description = "";
-      public int[] VariableIds = Array.Empty<int>();
-    }
 #nullable disable
 
-    private static ClusterRow[] ReadAllClustersFromTable(StringTable table)
+    private static ClusterExportRow[] ReadAllClustersFromTable(StringTable table)
     {
-      var list = new List<ClusterRow>();
+      var list = new List<ClusterExportRow>();
 
       foreach (var id in EnumerateIds(table, "Progesi.Cluster", "cluster:"))
       {
@@ -1461,7 +477,7 @@ CREATE TABLE IF NOT EXISTS ClusterVariables (
           desc,
           dto.Hashtag);
 
-        list.Add(new ClusterRow
+        list.Add(new ClusterExportRow
         {
           Id = id,
           Hash = cluster.Hashtag ?? "",
@@ -1475,9 +491,9 @@ CREATE TABLE IF NOT EXISTS ClusterVariables (
       return list.ToArray();
     }
 
-    private static VarRow[] ReadAllVarsFromTable(StringTable table)
+    private static VariableExportRow[] ReadAllVarsFromTable(StringTable table)
     {
-      var list = new List<VarRow>();
+      var list = new List<VariableExportRow>();
       foreach (var id in EnumerateIds(table, "Progesi.Var", "var:"))
       {
         string json = table.GetValue("Progesi.Var", "var:" + id.ToString(CultureInfo.InvariantCulture));
@@ -1487,54 +503,27 @@ CREATE TABLE IF NOT EXISTS ClusterVariables (
         if (dto == null) continue;
 
         object typed = ParseValue(dto.Value, dto.ValueType ?? "string");
-        string valc = ProgesiHash.CanonicalValue(typed);
         int[] deps = dto.Depends ?? Array.Empty<int>();
         bool ass = dto.IsAssumption ?? false;
-        string valueType = dto.ValueType ?? "string";
-        string excelValue;
-        string objectType = "";
-        string objectPayloadJson = "";
-        bool isExcelUnsupported;
-
-        if (typed is GeometryBase geometry)
-        {
-          objectType = ProgesiGeometryValueCodec.GetShortTypeName(geometry);
-          objectPayloadJson = ProgesiGeometryValueCodec.Encode(geometry);
-          excelValue = GhExcelObjectSheet.BuildObjectMarker(objectType);
-          isExcelUnsupported = false;
-        }
-        else
-        {
-          excelValue = GhExcelVariableValueSupport.FormatExportValue(valueType, dto.Value ?? "", valc);
-          isExcelUnsupported = GhExcelVariableValueSupport.RequiresUnsupportedExportHandling(valueType, dto.Value ?? "");
-        }
-
         var metadataIds = ReadVarDtoMetadataIds(dto);
-        var pv = new ProgesiVariable(id, dto.Name ?? "", typed ?? "", deps, metadataIds, ass);
-        string hash = ProgesiHash.Compute(pv);
 
-        list.Add(new VarRow
-        {
-          Id = id,
-          Hash = hash,
-          Name = dto.Name ?? "",
-          Value = excelValue,
-          ValC = valc,
-          MetaId = metadataIds.Length > 0 ? metadataIds[0] : 0,
-          MetadataIds = metadataIds,
-          Depends = deps,
-          Assumption = ass,
-          IsExcelUnsupported = isExcelUnsupported,
-          ObjectType = objectType,
-          ObjectPayloadJson = objectPayloadJson
-        });
+        list.Add(LiveExchangeVariableExportBuilder.Build(
+          id,
+          dto.Name ?? "",
+          typed,
+          dto.Value ?? "",
+          dto.ValueType ?? "string",
+          deps,
+          metadataIds,
+          ass,
+          GeometryCodec));
       }
       return list.ToArray();
     }
 
-    private static MetaRow[] ReadAllMetasFromTable(StringTable table)
+    private static MetadataExportRow[] ReadAllMetasFromTable(StringTable table)
     {
-      var list = new List<MetaRow>();
+      var list = new List<MetadataExportRow>();
       foreach (var id in EnumerateIds(table, "Progesi.Meta", "meta:"))
       {
         string json = table.GetValue("Progesi.Meta", "meta:" + id.ToString(CultureInfo.InvariantCulture));
@@ -1549,7 +538,7 @@ CREATE TABLE IF NOT EXISTS ClusterVariables (
         var meta = ProgesiMetadata.Create(by, desc, null, null, dto.LastModified, id);
         string hash = ProgesiHash.Compute(meta);
 
-        list.Add(new MetaRow
+        list.Add(new MetadataExportRow
         {
           Id = id,
           Hash = hash,
@@ -1559,7 +548,6 @@ CREATE TABLE IF NOT EXISTS ClusterVariables (
           LM = dto.LastModified.ToString("yyyy-MM-dd HH:mm:ss", CultureInfo.InvariantCulture)
         });
       }
-
 
       return list.ToArray();
     }
@@ -1581,45 +569,6 @@ CREATE TABLE IF NOT EXISTS ClusterVariables (
 
       ids.Sort();
       return ids.Distinct().ToArray();
-    }
-
-    private static List<GhExcelObjectSheet.ObjectChunkRow> ReadObjectChunkRows(XLWorkbook workbook)
-    {
-      var rows = new List<GhExcelObjectSheet.ObjectChunkRow>();
-      var wsO = GhExcelWorksheetLocator.TryGetWorksheet(
-        workbook,
-        GhExcelSheetNames.VariableObjects,
-        GhExcelSheetNames.VariableObjectsAlias);
-      if (wsO == null)
-        return rows;
-
-      var header = GhExcelHeaderMap.Build(wsO, out int headerRow, out int lastRow);
-      if (!header.TryGetValue("VARID", out int colVarId)
-          || !header.TryGetValue("CHUNKINDEX", out int colChunkIndex)
-          || !header.TryGetValue("CHUNKCOUNT", out int colChunkCount)
-          || !header.TryGetValue("OBJECTTYPE", out int colObjectType)
-          || !header.TryGetValue("PAYLOAD", out int colPayload))
-      {
-        return rows;
-      }
-
-      for (int r = headerRow + 1; r <= lastRow; r++)
-      {
-        int varId = GhExcelValueParsing.ToInt(wsO.Cell(r, colVarId).GetString());
-        if (varId <= 0)
-          continue;
-
-        rows.Add(new GhExcelObjectSheet.ObjectChunkRow
-        {
-          VarId = varId,
-          ChunkIndex = GhExcelValueParsing.ToInt(wsO.Cell(r, colChunkIndex).GetString()),
-          ChunkCount = GhExcelValueParsing.ToInt(wsO.Cell(r, colChunkCount).GetString()),
-          ObjectType = wsO.Cell(r, colObjectType).GetString() ?? "",
-          Payload = wsO.Cell(r, colPayload).GetString() ?? ""
-        });
-      }
-
-      return rows;
     }
 
     private static object ParseValue(string value, string valueType)
@@ -1663,45 +612,148 @@ CREATE TABLE IF NOT EXISTS ClusterVariables (
       return Array.Empty<int>();
     }
 
-    private static int[] ParseMetadataIds(string cellValue)
+    private sealed class RhinoGeometryValueCodecAdapter : IGeometryValueCodec
     {
-      if (string.IsNullOrWhiteSpace(cellValue))
-        return Array.Empty<int>();
+      public bool IsGeometryValueType(string valueType) =>
+        ProgesiGeometryValueCodec.IsGeometryValueType(valueType);
 
-      var trimmed = cellValue.Trim();
-      if (trimmed == "0")
-        return Array.Empty<int>();
-
-      if (!trimmed.Contains(",") && !trimmed.Contains(";") && !trimmed.Contains("|"))
+      public bool TryGetShortTypeName(object value, out string objectType)
       {
-        if (int.TryParse(trimmed, NumberStyles.Integer, CultureInfo.InvariantCulture, out var single) && single > 0)
-          return new[] { single };
-        return Array.Empty<int>();
+        if (value is GeometryBase geometry)
+        {
+          objectType = ProgesiGeometryValueCodec.GetShortTypeName(geometry);
+          return true;
+        }
+
+        objectType = null;
+        return false;
       }
 
-      var tokens = trimmed.Split(new[] { ',', ';', '|', ' ' }, StringSplitOptions.RemoveEmptyEntries);
-      var seen = new HashSet<int>();
-      var list = new List<int>();
-      foreach (var token in tokens)
+      public bool TryEncode(object value, out string objectType, out string payloadJson)
       {
-        if (int.TryParse(token.Trim(), NumberStyles.Integer, CultureInfo.InvariantCulture, out var n)
-            && n > 0
-            && seen.Add(n))
-          list.Add(n);
+        if (value is GeometryBase geometry)
+        {
+          objectType = ProgesiGeometryValueCodec.GetShortTypeName(geometry);
+          payloadJson = ProgesiGeometryValueCodec.Encode(geometry);
+          return true;
+        }
+
+        objectType = null;
+        payloadJson = null;
+        return false;
       }
 
-      return list.ToArray();
+      public bool TryDecode(string payloadJson, out object geometry)
+      {
+        if (ProgesiGeometryValueCodec.TryDecode(payloadJson, out var decoded) && decoded != null)
+        {
+          geometry = decoded;
+          return true;
+        }
+
+        geometry = null;
+        return false;
+      }
     }
 
-    private static string FormatMetaIdForExcel(int[] metadataIds)
+    private sealed class RhinoLiveExchangeImportSink : ILiveExchangeImportSink
     {
-      if (metadataIds == null || metadataIds.Length == 0)
-        return string.Empty;
+      private readonly object _repo;
+      private readonly StringTable _clusterTable;
+      private int _maxMetaId;
+      private int _maxVarId;
 
-      if (metadataIds.Length == 1)
-        return metadataIds[0].ToString(CultureInfo.InvariantCulture);
+      public RhinoLiveExchangeImportSink(object repo, StringTable clusterTable)
+      {
+        _repo = repo;
+        _clusterTable = clusterTable;
+      }
 
-      return string.Join(",", metadataIds);
+      public bool TryUpsertMetadata(int id, string by, string description, string refsPipeSeparated, out int persistedId, out string error)
+      {
+        persistedId = 0;
+        var payload = new { id, by, info = description, rf = refsPipeSeparated, sn = "" };
+        object persisted;
+        bool ok = MetadataRepositoryCompatExtensions.TryUpsert(_repo, payload, out persisted, out error);
+        if (ok)
+        {
+          ReadPersistedId(persisted, ref persistedId);
+          if (persistedId > _maxMetaId) _maxMetaId = persistedId;
+        }
+
+        return ok;
+      }
+
+      public bool TryGetMetadataById(int id, out string error)
+      {
+        object dummy;
+        return MetadataRepositoryCompatExtensions.TryGetByHashThenId(_repo, "", id, out dummy, out error);
+      }
+
+      public bool TryUpsertVariable(LiveExchangeVariableImportPayload payload, out int persistedId, out string error)
+      {
+        persistedId = 0;
+        var repoPayload = new
+        {
+          id = payload.Id,
+          name = payload.Name,
+          value = payload.GeometryJson ?? payload.Value,
+          geometryJson = payload.GeometryJson ?? "",
+          unit = "",
+          by = "",
+          isAssumption = payload.IsAssumption,
+          metadataIds = (object)payload.MetadataIds,
+          mid = LiveExchangeMetadataIds.FormatForExcel(payload.MetadataIds),
+          depends = (object)payload.Depends
+        };
+
+        object persisted;
+        bool ok = MetadataRepositoryCompatExtensions.TryUpsertVariable(_repo, repoPayload, out persisted, out error);
+        if (ok)
+        {
+          ReadPersistedId(persisted, ref persistedId);
+          if (persistedId > _maxVarId) _maxVarId = persistedId;
+        }
+
+        return ok;
+      }
+
+      public bool TryPersistCluster(LiveExchangeClusterImportPayload cluster)
+      {
+        if (_clusterTable == null) return false;
+
+        var dto = new ClusterDto
+        {
+          Id = cluster.Id,
+          Name = cluster.Name,
+          Description = cluster.Description,
+          VariableIds = cluster.VariableIds,
+          Hashtag = cluster.Hashtag
+        };
+
+        string json = JsonConvert.SerializeObject(dto);
+        _clusterTable.SetString("Progesi.Cluster", "cluster:" + cluster.Id.ToString(CultureInfo.InvariantCulture), json);
+
+        int next = ReadCounter(_clusterTable, "Progesi.Cluster");
+        if (cluster.Id + 1 > next)
+          _clusterTable.SetString("Progesi.Cluster", "__next__", (cluster.Id + 1).ToString(CultureInfo.InvariantCulture));
+
+        return true;
+      }
+
+      public void UpdateIdCounters(int maxMetaId, int maxVarId)
+      {
+        var doc = RhinoDoc.ActiveDoc ?? throw new InvalidOperationException("RhinoDoc.ActiveDoc is null.");
+        StringTable table = doc.Strings ?? throw new InvalidOperationException("RhinoDoc.Strings is null.");
+        int curMetaNext = ReadCounter(table, "Progesi.Meta");
+        int curVarNext = ReadCounter(table, "Progesi.Var");
+        int metaTarget = Math.Max(maxMetaId, _maxMetaId);
+        int varTarget = Math.Max(maxVarId, _maxVarId);
+        if (metaTarget > 0)
+          table.SetString("Progesi.Meta", "__next__", Math.Max(curMetaNext, metaTarget + 1).ToString(CultureInfo.InvariantCulture));
+        if (varTarget > 0)
+          table.SetString("Progesi.Var", "__next__", Math.Max(curVarNext, varTarget + 1).ToString(CultureInfo.InvariantCulture));
+      }
     }
   }
